feat: 별점(평가) api 연동 진행중

+++ app/src/main/java/com/beside153/peopleinside/model/contentdetail/ContentRatingRequest.kt
@@ -8,5 +8,5 @@ import kotlinx.serialization.Serializable
 @Parcelize
 @Serializable
 data class ContentRatingRequest(
-    @SerialName("rating") val rating: Double
+    @SerialName("rating") val rating: Float
 ) : Parcelable
